Develop

+++ groomeet_backend/bandas.py
@@ -3,13 +3,18 @@ from django.shortcuts import redirect, render, get_object_or_404
 from groomeet_backend.form import *
 from django.contrib.auth.decorators import login_required
 from django.contrib import messages
+import os
 
-@login_required
+@login_required(login_url='/login/')
 def bandaCreate(request):
     if request.method == "POST":
         formulario = BandaForm(request.POST)
         if formulario.is_valid():
-            banda = Banda.objects.create(nombre = formulario.cleaned_data['nombre'], administrador=Musico(request.user.pk))
+            banda = Banda.objects.create(nombre = formulario.cleaned_data['nombre'], administrador=get_object_or_404(Musico,usuario = request.user),
+                                        descripcion=formulario.cleaned_data['descripcion'],enlaceVideo = formulario.cleaned_data['enlaceVideo'])
+            if 'imagen' in request.FILES:
+                banda.imagen = imagen=request.FILES['imagen']
+            banda.save()
             banda.generos.set(request.POST.getlist('generos'))
             banda.instrumentos.set(request.POST.getlist('instrumentos'))
             
@@ -18,25 +23,48 @@ def bandaCreate(request):
         formulario = BandaForm()
     return render(request, 'createBanda.html', {'formulario': formulario})
 
-@login_required
+@login_required(login_url='/login/')
 def bandaUpdate(request, id):
-    banda = Banda.objects.get(id=id)
-    form = BandaForm(initial={'nombre': banda.nombre})
+    banda = get_object_or_404(Banda,id=id)
+    musico = get_object_or_404(Musico,usuario=request.user)
+    if banda.administrador != musico:
+        return redirect('/misBandas')
+    imagenBanda = banda.imagen
+    form = BandaForm(initial={'nombre': banda.nombre,'descripcion': banda.descripcion,'instrumentos': banda.instrumentos.all,
+                            'generos': banda.generos.all,'imagen': banda.imagen,'enlaceVideo': banda.enlaceVideo})
     if request.method == "POST":
         form = BandaForm(request.POST, instance=banda)
         if form.is_valid():
             try:
                 form.save()
                 model = form.instance
+                #Si existe una nueva imagen, habrá que cambiar la existente por esta.
+                if 'imagen' in request.FILES:
+                    print("He recibido la imagen")
+                    #Borramos la imagen anterior si existia.
+                    if imagenBanda != "" and imagenBanda != None:
+                        os.remove(banda.imagen.path)
+                    model.imagen = request.FILES['imagen']
+                    model.save()
+                else:
+                    print("Nada bro, no He recibido la imagen")
+                #En caso contrario, comprobaremos que el usuario no ha eliminado la imagen que tenía.
+                #Si la ha eliminado habrá que borrarla de los datos.
+                    if model.imagen == "" or model.imagen == None:
+                        os.remove(imagenBanda.path)
+                messages.success(request, f"¡Tu banda ha sido modificada con éxito!")
                 return redirect('/misBandas')
             except Exception as e:
                 pass
     return render(request, 'updateBanda.html', {'form': form})
 
 
-@login_required
+@login_required(login_url='/login/')
 def bandaDelete(request, id):
-    banda = Banda.objects.get(id=id)
+    banda = get_object_or_404(Banda,id=id)
+    musico = get_object_or_404(Musico,usuario=request.user)
+    if banda.administrador != musico:
+        return redirect('/misBandas')
     try:
         banda.delete()
     except:
@@ -44,11 +72,12 @@ def bandaDelete(request, id):
     return redirect('/misBandas')
 
 
-@login_required
+@login_required(login_url='/login/')
 def miembroNoRegistradoCreate(request, pk):
-    banda = get_object_or_404(Banda, id=pk)
-    if banda.administrador.id != request.user.id:
-        return HttpResponseRedirect('/misBandas')
+    banda = get_object_or_404(Banda,id=pk)
+    musico = get_object_or_404(Musico,usuario=request.user)
+    if banda.administrador != musico:
+        return redirect('/misBandas')
     if request.method == "POST":
         formulario = MiembroNoRegistradoForm(request.POST)
         if formulario.is_valid():
@@ -61,7 +90,70 @@ def miembroNoRegistradoCreate(request, pk):
         formulario = MiembroNoRegistradoForm()
     return render(request, 'createMiembroNoRegistrado.html', {'formulario': formulario})
 
-@login_required
+@login_required(login_url='/login/')
+def enviarInvitacionBanda(request, banda_id):
+    if request.method == "POST":
+        formulario = InvitarBandaForm(request.POST)
+        if formulario.is_valid():
+            estado = EstadoInvitacion.Pendiente
+            usuario = get_object_or_404(User, username=formulario.cleaned_data['receptor'])
+            banda = get_object_or_404(Banda, id=banda_id)
+
+            #Comprobando que el usuario no pertenece ya a la banda
+            try:
+                invitacion_aceptada = Invitacion.objects.get(receptor = receptor, 
+                                                    banda = banda, estado = EstadoInvitacion.Aceptada)
+                messages.error = (request, f"El usuario {usuario.username} ya pertenece a la banda {banda.nombre}")
+            except:
+                invitacion_aceptada = None
+
+            #Comprobando que el usuario no tiene ya una invitación pendiente para esa banda
+            try:
+                invitacion_pendiente = Invitacion.objects.get(receptor = receptor, 
+                                                    banda = banda, estado = EstadoInvitacion.Pendiente)
+                messages.error = (request, f"El usuario {usuario.username} ya tiene una invitación pendiente para la banda {banda.nombre}")
+            except:
+                invitacion_pendiente = None
+
+            if invitacion_aceptada == None and invitacion_pendiente == None:
+                invitacion = Invitacion.objects.create(receptor = get_object_or_404(Musico,usuario=usuario), emisor=get_object_or_404(Musico,usuario = request.user),
+                                                banda=banda, estado=estado)
+                messages.success = (request, f"¡La invitación a {usuario.username} para la banda {banda.nombre} fue enviada!")
+            return HttpResponseRedirect('/misBandas')
+    else:
+        formulario = InvitarBandaForm()
+    return render(request, 'invitarBanda.html', {'formulario': formulario})
+
+
+@login_required(login_url='/login/')
+def aceptarInvitacionBanda(request, invitacion_id):
+    usuario = request.user
+    receptor = get_object_or_404(Musico, usuario=usuario)
+    invitacion = get_object_or_404(Invitacion, id=invitacion_id, receptor = receptor, estado = EstadoInvitacion.Pendiente)
+
+    banda = invitacion.banda
+    nuevo_miembro = MiembroDe.objects.create(musico = receptor, banda = banda)
+    invitacion.estado = EstadoInvitacion.Aceptada
+    invitacion.save()
+    messages.success = (request, f"¡Te has unido a la banda {banda.nombre}!")
+
+    return redirect("/misInvitaciones")
+
+@login_required(login_url='/login/')
+def rechazarInvitacionBanda(request, invitacion_id):
+    usuario = request.user
+    receptor = get_object_or_404(Musico, usuario=usuario)
+    invitacion = get_object_or_404(Invitacion, id=invitacion_id, receptor = receptor, estado = EstadoInvitacion.Pendiente)
+
+    invitacion.estado = EstadoInvitacion.Rechazada
+    invitacion.save()
+    messages.success = (request, f"¡Has rechazado la invitación a la banda {invitacion.banda.nombre}!")
+
+    return redirect("/misInvitaciones")
+
+#Este sera el método utilizado para cuando se implemente las invitaciones en el propio chat
+'''
+@login_required(login_url='/login/')
 def enviarInvitacionBanda(request, receptor_id, banda_id):
     emisor = get_object_or_404(Musico, usuario=request.user)
     receptor = get_object_or_404(Musico, id=receptor_id)
@@ -94,22 +186,4 @@ def enviarInvitacionBanda(request, receptor_id, banda_id):
             messages.error = (request, f"La invitación no se pudo enviar")
         
     return redirect("/listado")
-
-@login_required
-def aceptarInvitacionBanda(request, banda_id):
-    usuario = request.user
-    receptor = get_object_or_404(Musico, usuario=usuario)
-    banda = get_object_or_404(Banda, id=banda_id)
-    try:
-        invitacion = Invitacion.objects.get(receptor = receptor, banda = banda, estado = EstadoInvitacion.Pendiente)
-    except:
-        invitacion = None
-        messages.error = (request, f"La invitación no existe")
-
-    if invitacion != None:
-        nuevo_miembro = MiembroDe.objects.create(musico = receptor, banda = banda)
-        invitacion.estado = EstadoInvitacion.Aceptada
-        invitacion.save()
-        messages.success = (request, f"¡Te has unido a la banda {banda.nombre}!")
-
-    return redirect("/listado")
+'''
